requirements.txt
Flask==3.0.0
langchain==0.1.0
langchain-community==0.0.10
pypdf2==3.0.1
pdfplumber==0.10.3
faiss-cpu==1.7.4
sentence-transformers>=3.0.1
transformers>=4.44.0,<5
huggingface-hub>=0.34.0,<1.0
python-dotenv==1.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
chromadb==0.4.24
docarray==0.40.0
langchain-openai>=0.0.2
